Added shipping-dollar and shipping-cents
Formerly we just had one product, "shipping". Now there is one for
pennies and one for dollars, which is a little less crazy. We might
even want to add a third later, for 10's or another for 100's!

I modified things so that now all we need to do is add a new product
to the shipping collection called 'Shipping-denomination'.